Multi Group implementation + migration

+++ protected/humhub/modules/admin/controllers/GroupController.php
@@ -12,7 +12,8 @@ use Yii;
 use yii\helpers\Url;
 use humhub\modules\admin\components\Controller;
 use humhub\modules\user\models\Group;
-use humhub\modules\user\models\UserFilter;
+use humhub\modules\user\widgets\UserPicker;
+use humhub\modules\user\models\User;
 
 /**
  * Group Administration Controller
@@ -53,16 +54,18 @@ class GroupController extends Controller
             $group = new Group();
         }
 
-        $group->scenario = 'edit';
+        $group->scenario = Group::SCENARIO_EDIT;
         $group->populateDefaultSpaceGuid();
         $group->populateAdminGuids();
          
+        
+        
         if ($group->load(Yii::$app->request->post()) && $group->validate()) {
             $group->save();
             $this->redirect(Url::toRoute('/admin/group'));
         }
 
-        $showDeleteButton = (!$group->isNewRecord && Group::find()->count() > 1);
+        $showDeleteButton = (!$group->isNewRecord && !$group->is_admin_group);
 
         // Save changed permission states
         if (!$group->isNewRecord && Yii::$app->request->post('dropDownColumnSubmit')) {
@@ -89,28 +92,31 @@ class GroupController extends Controller
     public function actionDelete()
     {
         $group = Group::findOne(['id' => Yii::$app->request->get('id')]);
+        
         if ($group == null) {
             throw new \yii\web\HttpException(404, Yii::t('AdminModule.controllers_GroupController', 'Group not found!'));
-        }
-
-        //TODO: obsolete ? foreignkeys
-        foreach($group->groupUsers as $groupUser) {
-            $groupUser->delete();
         } 
-        $group->delete();
+        
+        //Double check to get sure we don't remove the admin group
+        if(!$group->is_admin_group) {
+            $group->delete();
+        }
+        
         $this->redirect(Url::toRoute("/admin/group"));
     }
     
     public function actionAdminUserSearch()
     {
         Yii::$app->response->format = 'json';
-        $maxResult = 10;
+        
         $keyword = Yii::$app->request->get('keyword');
-        
         $group = Group::findOne(Yii::$app->request->get('id'));
-        $user = UserFilter::filter($group->getUsers(), $keyword, $maxResult);
         
-        return \humhub\modules\user\widgets\UserPicker::asJSON($user);
+        return UserPicker::filter([
+            'query' => $group->getUsers(),
+            'keyword' => $keyword,
+            'fillQuery' => User::find()
+        ]);
     }
 
 }
